+++ lib/src/browser_theme_color_stub.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+
+class BrowserThemeColorSync extends StatelessWidget {
+  const BrowserThemeColorSync({
+    super.key,
+    required this.brightness,
+    required this.child,
+  });
+
+  final Brightness brightness;
+  final Widget child;
+
+  @override
+  Widget build(BuildContext context) => child;
+}